Deleted old FaceRecognition scripts and updated face_recognition_rtsp_v1.py
