+++ FaceRecognition/face_recognition_rtsp_v1.py
@@ -6,11 +6,12 @@ import threading
 
 # Global constants and shared variables
 FACEBANK_PATH = "facebank"
-THRESHOLD = 0.45  # Face distance threshold for recognition
+THRESHOLD = 0.5  # Face distance threshold for recognition
 
 # Shared dictionary for the latest processed frame from each stream
 latest_frames = {}
 frame_lock = threading.Lock()  # Protects access to latest_frames
+
 
 def compute_confidence(distance, threshold=THRESHOLD, k=10):
     """
@@ -20,10 +21,55 @@ def compute_confidence(distance, threshold=THRESHOLD, k=10):
     """
     confidence = 1 / (1 + np.exp(k * (distance - threshold))) * 100
     return confidence
+
+
+def combine_frames_grid(frames, cols=2):
+    """
+    Combine a list of frames into a grid layout with the given number of columns.
+    The frames are resized so that they all have the same height.
     
+    Args:
+        frames: List of frames (numpy arrays).
+        cols: Number of columns in the grid.
+        
+    Returns:
+        A single image with all frames arranged in a grid.
+    """
+    if not frames:
+        return None
+
+    # Resize frames to the same height.
+    # Use the minimum height among the frames.
+    min_height = min(frame.shape[0] for frame in frames)
+    resized_frames = []
+    for frame in frames:
+        h, w = frame.shape[:2]
+        scale = min_height / h
+        new_w = int(w * scale)
+        resized_frames.append(cv2.resize(frame, (new_w, min_height)))
+
+    # Organize frames into rows.
+    rows = []
+    for i in range(0, len(resized_frames), cols):
+        # Extract one row of frames.
+        row_frames = resized_frames[i:i + cols]
+        # If the row has fewer frames than the desired columns,
+        # you might want to pad with black images.
+        if len(row_frames) < cols:
+            # Create a black image with same height and width equal to the first frame's width.
+            black_frame = np.zeros_like(row_frames[0])
+            row_frames.extend([black_frame] * (cols - len(row_frames)))
+        row = np.hstack(row_frames)
+        rows.append(row)
+
+    # Stack all rows vertically.
+    grid = np.vstack(rows)
+    return grid
+
 
 class FaceRecognition:
     """Handles loading the facebank and performing face recognition."""
+
     def __init__(self, facebank_path):
         self.facebank = self.load_facebank(facebank_path)
 
@@ -39,8 +85,7 @@ class FaceRecognition:
                     "stevenle": "Steven Le",
                     "capham": "Ca Pham"
                 }
-                name = name_map.get(base_name, base_name)  # Fallback if not in mapping
-
+                name = name_map.get(base_name, base_name)
                 image_path = os.path.join(path, filename)
                 image = face_recognition.load_image_file(image_path)
                 encodings = face_recognition.face_encodings(image)
@@ -51,14 +96,16 @@ class FaceRecognition:
         return facebank
 
     def recognize_faces(self, frame):
-        """Detects faces in a frame and returns their locations and recognized names."""
+        """
+        Detects faces in a frame and returns their locations and recognized names.
+        The frame provided here can be a cropped ROI.
+        """
         rgb_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
         face_locations = face_recognition.face_locations(rgb_frame, model="cnn")
         face_encodings = face_recognition.face_encodings(rgb_frame, face_locations)
         results = []
-
         for (top, right, bottom, left), face_encoding in zip(face_locations, face_encodings):
-            label = "Not Registerd"
+            label = "Not Registered"
             distances = [
                 (bank_name, face_recognition.face_distance([bank_encoding], face_encoding)[0])
                 for bank_name, bank_encoding in self.facebank.items()
@@ -73,44 +120,73 @@ class FaceRecognition:
 
 
 class RTSPStream:
-    """Manages one RTSP stream and processes frames for face recognition."""
+    """
+    Manages one RTSP stream and processes frames for face recognition,
+    liveness check, and (optionally) a Region of Interest (ROI).
+    """
+
     def __init__(self, name, rtsp_url, face_recognizer, roi=None):
+        """
+        roi: Optional tuple (x, y, width, height) specifying the ROI for face recognition.
+        """
         self.name = name
         self.rtsp_url = rtsp_url
         self.face_recognizer = face_recognizer
-        self.roi = roi
+        self.roi = roi  # For example, (0, 0, frame_width, 200)
         self.video_capture = self.create_capture_pipeline(rtsp_url)
         self.running = True
+        self.prev_frame = None
 
     def create_capture_pipeline(self, rtsp_url):
-        """Create an OpenCV VideoCapture using a GStreamer pipeline."""
-        pipeline = (
-            f"rtspsrc location={rtsp_url} latency=100 protocols=udp drop-on-latency=true ! "
-            "rtph264depay ! h264parse ! queue max-size-buffers=1 ! nvv4l2decoder ! "
-            "nvvidconv ! videoconvert ! video/x-raw,format=BGR ! "
-            "appsink drop=true max-buffers=1 sync=false"
-        )
+        """
+        Depending on the URL, use a different GStreamer pipeline.
+        For the third stream (if "aircity2025.
+        """
+        if "aircity2025" in rtsp_url:
+            # Use a simpler pipeline with avdec_h264 (and ensure credentials are URL-encoded)
+            pipeline = (
+                f"rtspsrc location={rtsp_url} latency=100 ! "
+                "rtph264depay ! h264parse ! avdec_h264 ! "
+                "videoconvert ! video/x-raw,format=BGR,width=640,height=480 ! "
+                "appsink drop=true max-buffers=1 sync=false "
+            )
+        else:
+            pipeline = (
+                f"rtspsrc location={rtsp_url} latency=100 protocols=udp drop-on-latency=true ! "
+                "rtph264depay ! h264parse ! queue max-size-buffers=1 ! nvv4l2decoder ! "
+                "nvvidconv ! videoconvert ! video/x-raw,format=BGR,width=640,height=480 ! "
+                "appsink drop=true max-buffers=1 sync=false"
+            )
         capture = cv2.VideoCapture(pipeline, cv2.CAP_GSTREAMER)
         if not capture.isOpened():
             print(f"[ERROR] Unable to open RTSP stream: {rtsp_url}")
         return capture
 
     def process_stream(self):
-        """Continuously reads frames, performs face recognition, and updates the shared frame dictionary."""
+        """
+        Continuously reads frames, performs liveness check, and recognizes faces
+        only in the defined ROI (if provided). Also overlays the camera name on the frame.
+        """
+        first_warning = True
         while self.running:
             ret, frame = self.video_capture.read()
             if not ret:
-                print(f"[WARNING] Failed to grab frame from {self.rtsp_url}")
-                continue
+                if first_warning:
+                    print(f"[WARNING] Failed to grab frame from {self.rtsp_url}")
+                    first_warning = False
+                # Create a black frame if the stream is not running
+                frame = np.zeros((480, 640, 3), dtype=np.uint8)
+                cv2.putText(frame, "Stream Not Available", (10, 30), cv2.FONT_HERSHEY_SIMPLEX,
+                            1, (0, 0, 255), 2)
 
-            # Process the frame for face recognition
+            # If ROI is defined, use that portion for face recognition.
             if self.roi is not None:
                 (roi_x, roi_y, roi_w, roi_h) = self.roi
-                # Draw the ROI rectangle on the full frame for visualization
+                # Draw the ROI rectangle on the full frame for visualization.
                 cv2.rectangle(frame, (roi_x, roi_y), (roi_x + roi_w, roi_y + roi_h), (255, 0, 0), 2)
-                roi_frame = frame[roi_y:roi_y+roi_h, roi_x:roi_x+roi_w]
+                roi_frame = frame[roi_y:roi_y + roi_h, roi_x:roi_x + roi_w]
                 faces = self.face_recognizer.recognize_faces(roi_frame)
-                # Adjust the face coordinates back to the original frame
+                # Adjust detected face coordinates back to the full frame.
                 adjusted_faces = []
                 for (top, right, bottom, left, label) in faces:
                     adjusted_faces.append((top + roi_y, right + roi_x, bottom + roi_y, left + roi_x, label))
@@ -118,20 +194,18 @@ class RTSPStream:
             else:
                 faces = self.face_recognizer.recognize_faces(frame)
 
-
-            for (top, right, bottom, left, name) in faces:
+            for (top, right, bottom, left, label) in faces:
                 cv2.rectangle(frame, (left, top), (right, bottom), (0, 255, 0), 2)
-                cv2.putText(frame, name, (left, top - 10), cv2.FONT_HERSHEY_SIMPLEX,
+                cv2.putText(frame, label, (left, top - 10), cv2.FONT_HERSHEY_SIMPLEX,
                             0.9, (255, 255, 255), 2)
-            
-            # Overlay the camera name on the frame (e.g., top-left corner)
+
+            # Overlay the camera name on the frame (top-left corner).
             cv2.putText(frame, self.name, (10, 30), cv2.FONT_HERSHEY_SIMPLEX,
                         1, (0, 255, 255), 2)
 
-
-            # Update the shared latest_frames dictionary
             with frame_lock:
                 latest_frames[self.name] = frame.copy()
+            self.prev_frame = frame.copy()
 
         self.video_capture.release()
 
@@ -141,22 +215,18 @@ def main():
     rtsp_streams = {
         "Labs": "rtsp://admin:L2F2A85E@192.168.1.192:554/cam/realmonitor?channel=1&subtype=1",
         "Spaceship": "rtsp://admin:L297FC1C@192.168.1.185:554/cam/realmonitor?channel=1&subtype=1",
-        "Blue wall": "rtsp://admin:@dm1ncamera@luxypq3.duckdns.org:557/Streaming/channels/802"
+        "HIK Vision 1": "rtsp://admin:aircity2025@192.168.1.2:554/Streaming/channels/202"
     }
-    
 
     # Define a dictionary mapping camera names to ROI tuples.
     # For example, here we set an ROI for "Camera Labs" (top of the frame) and leave others as full frame.
     camera_rois = {
-        "Labs": (50,30,400,300),
-        "Spaceship": (10,100,200,200),
-        "Blue Wall": None  # No ROI defined for Camera 3.
+        "Labs": (50, 100, 500, 350),
+        "Spaceship": (20, 250, 400, 200),
+        "HIK Vision 1": (10, 10, 600, 400)  # No ROI defined for Camera 3.
     }
 
-    # Initialize face recognition using the facebank
     face_recognizer = FaceRecognition(FACEBANK_PATH)
-
-    # Start each RTSP stream in its own thread
     streams = []
     threads = []
     for name, url in rtsp_streams.items():
@@ -167,32 +237,26 @@ def main():
         threads.append(thread)
         thread.start()
 
-    # Main loop to combine frames from all streams and display them in one window
     while True:
         with frame_lock:
-            # Only use frames that have been received
             frames = [frame for frame in latest_frames.values() if frame is not None]
-
         if frames:
             try:
-                # Combine frames horizontally (ensure they have the same height)
-                combined_frame = np.hstack(frames)
-                cv2.imshow("Combined Cameras", combined_frame)
+                combined_frame = combine_frames_grid(frames, cols=3)
+                cv2.imshow("AI Security Camera Streams", combined_frame)
             except Exception as e:
                 print("Error combining frames:", e)
-
-        # Press 'q' to exit
         if cv2.waitKey(1) & 0xFF == ord('q'):
             break
 
-    # Stop all stream threads gracefully
     for stream in streams:
         stream.running = False
     for t in threads:
         t.join()
-
     cv2.destroyAllWindows()
 
 
 if __name__ == "__main__":
     main()
+
+# happy code
